Refactor and redesign subscribe dialog
- Move Subscribe dialog to other dialogs
- Redesign Sub Dialog to fit the other dialogs
